+++ packages/newspringchurchapp/src/prayer/AddPrayer/AddPrayerForm/AddPrayerForm.tests.js
@@ -2,7 +2,7 @@ import React from 'react';
 import renderer from 'react-test-renderer';
 import Providers from 'newspringchurchapp/src/Providers';
 
-import AddPrayerForm from '.';
+import AddPrayerForm from './AddPrayerForm';
 
 describe('The AddPrayerForm component', () => {
   it('should render', () => {
