+++ packages/newspringchurchapp/src/prayer/getProfilePic.js
@@ -1,0 +1,14 @@
+import gql from 'graphql-tag';
+
+export default gql`
+  query {
+    currentUser {
+      id
+      profile {
+        photo {
+          uri
+        }
+      }
+    }
+  }
+`;
