+++ packages/newspringchurchapp/src/tabs/prayer/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { createStackNavigator } from 'react-navigation';
 import { tabBarIconNewSpring } from '../tabBarIcon';
-import AddPrayerForm from '../../prayer/AddPrayer/AddPrayerForm';
+import { AddPrayerFormConnected } from '../../prayer/AddPrayer/AddPrayerForm';
 import Prayer from './Prayer';
 
 const PrayerNavigator = createStackNavigator(
@@ -9,7 +9,7 @@ const PrayerNavigator = createStackNavigator(
     Prayer,
     AddPrayerForm: {
       // TODO: remove the lint disable once react-navigation can work well with memo
-      screen: () => <AddPrayerForm />, // eslint-disable-line react/display-name
+      screen: () => <AddPrayerFormConnected />, // eslint-disable-line react/display-name
       navigationOptions: { header: null },
     },
   },
